go back to annoying workaround, again
