+++ mealie/db/models/_model_base.py
@@ -1,11 +1,11 @@
 import string
 from datetime import datetime
-from typing import Annotated
 
 from sqlalchemy import Integer
 from sqlalchemy.orm import DeclarativeBase, Mapped, declared_attr, mapped_column, synonym
 from text_unidecode import unidecode
 
+from ._filterable_column import FilterableColumn
 from ._model_utils.datetime import NaiveDateTime, get_utc_now
 
 # Punctuation characters replaced with spaces during text normalization.
@@ -13,18 +13,6 @@ from ._model_utils.datetime import NaiveDateTime, get_utc_now
 # double-quote, which are reserved for quoted literal searches.
 NORMALIZE_PUNCTUATION = string.punctuation.replace("'", "").replace('"', "")
 _NORMALIZE_PUNCTUATION_TABLE = str.maketrans(NORMALIZE_PUNCTUATION, " " * len(NORMALIZE_PUNCTUATION))
-
-
-class FilterableColumn[T](Mapped):
-    """
-    Drop-in replacement for `Mapped[]` that marks a column as filterable.
-    Filterable columns can be used in query filter expressions.
-
-    Only valid on scalar column fields. Using it on a relationship type (e.g. `list[Model]`).
-    """
-
-    def __class_getitem__(cls, item: type) -> type:
-        return Mapped[Annotated[item, mapped_column(info={"filterable": True})]]
 
 
 class SqlAlchemyBase(DeclarativeBase):
